Initial prototype for Entry class.
This is the first version of Entry, which does not yet support containing
vectors. Entry exists to allow a single map to contain any value type, so
that only a single map<string, Entry> can be used in place of different maps
for each type of value that the DataBlock's sections must be able to hold.