Out-of-source build now working for Fortran, but without dependency generation.
Automatic dependency generation for Fortran modules is a difficult problem. If we
really want to support it, we may want to change our build system to one that uses
either CMake or SCons, both of which seem to have solved this problem.

For now, users are responsible for writing their own dependencies. An example is the
contents of cosmosis/src/datablock/cosmosis_f90/Makefile, which compiles several Fortran
source code files which contain modules with dependencies between them, and then links
the resulting .o files into a dynamic library.