First step for full in-source build: cosmosis/cosmosis/datablock done
This is just the first step. From the top-level cloned repository,
'make all' recurses into subdirectories, and builds whatever the
Makefiles in the entered directories says to build. Note that targets
must be in place in each of these makefiles, and that those targets
must kick off the recursion again, if necessary.

Currently 'all', 'clean', and 'test' targets are supported.