Use full path name for creating dynamic libraries.
By using the full path name to create dynamic libraries, we tell
the Apple linker to set its internal LC_ID_DYLIB to be the full
path to the library. This allows it to be found even on a system
using SIP. For the GNU linker, this just makes the output file
with the given name.